+++ ShipMonitor/ShipTransferInitiatedEvent.cs
@@ -43,7 +43,7 @@ namespace EddiShipMonitor
 
         public ShipTransferInitiatedEvent(DateTime timestamp, string ship, int? shipid, string system, decimal distance, long? price, long? time) : base(timestamp, NAME)
         {
-            this.ship = ship;
+            this.ship = ShipDefinitions.FromEDModel(ship).model;
             this.shipid = shipid;
             this.system = system;
             this.distance = distance;